+++ NEWS.md
@@ -1,12 +1,15 @@
-# emuR 2.1.0.9001
+# emuR 2.1.0.9002
 
 ## new features / performance tweaks / improvements
+
+- `load_emuDB()` performance bump due to SQLite transaction improvement + using prep. statements 
 
 ## bug fixes
 
 - implemented first version of `delete_itemsInLevel()`
 - `requery_seq()` now returns correct attribute label vs. the level attribute
 - fixed `In rbind(names(params), unlist(params, recursive = F)): number of columns of result is not a multiple of vector length (arg 1)` warning in `runBASwebservice_maus()` with `httr::upload_file()`
+- added missing `()` to `return` statement
 
 # emuR 2.1.0